requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
foursquare==1!2016.9.12
idna==2.8
numpy==1.15.4
pandas==0.23.4
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.7
requests==2.21.0
six==1.12.0
urllib3==1.24.1
